# BackEnd/Models/WorkersApsi.cs
using System;
using System.Collections.Generic;
using System.ComponentModel.DataAnnotations;
using System.Linq;
using System.Threading.Tasks;

namespace BackEnd.Models
{
    public class WorkersApsi
    {
        public int Id { get; set; }
        [Required]
        public String Name { get; set; }
        [Required]
        public String LastName { get; set; }
        [Required]
        public String DocId { get; set; }
        [Required]
        public String Job { get; set; }
    }
}